Suppress warnings when using deprecated APIs
When the usage is intentional. These can be grepped for and reviewed
again later.

This is not possible for the Camlp4 extension, probably due to

  https://github.com/ocaml/camlp4/issues/55

since Camlp4 runs on the source of the extension. Will be dealt with in
a future commit that enables/disables warnings on a per-file basis.